업로드 파일 dto를 하나로 합쳐라
업로드 파일 dto를 하나로 합칩니다
